Calculator Functional
The calculator is now entirely functional for the four basic operations,
though the decimal does not work yet. Large numbers currently go off of
the screen, so some form of rounding or other constraints will be
necessary to keep the numbers on the screen.

Other future considerations:
-Currently no way to clear the screen
-Currently no way to tell what was previously entered. Will need to
subdivide the screen in order to capture this other information
-May want to include other functionality, such as powers, square root
and more

[Ticket:x]